+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [1.0-alpha13] - 2022-09-22
+
 ### Changed
 
 - `BookAppointmentActivity` has been renamed to `ScheduleAppointmentActivity`.